Calc mark & make response JSON

+++ src/main/java/com/bionic/entities/Result.java
@@ -71,7 +71,7 @@ public class Result {
         this.submited = submited;
     }
 
-    public float getMark() {
+    public int  getMark() {
         return mark;
     }
